[Add] keep alive method

+++ Client.h
@@ -1,4 +1,4 @@
 #pragma once
 
 int client_start();
-UINT16 client_initialize_connection(long rc, SOCKET s, SOCKADDR_IN addr, SOCKADDR_IN* remote_addr, int remote_addr_len, UINT16 msg_size);
+UINT16 client_initialize_connection(long rc, SOCKET s, SOCKADDR_IN addr, SOCKADDR_IN* remote_addr, int remote_addr_len, UINT32 msg_size);
